feat(vrack-services): use MRC layouts and utils
ref: #MANAGER-16779
