+++ packages/manager/apps/vrack-services/__mocks__/iam/index.ts
@@ -1,0 +1,1 @@
+export * from './iam';
