+++ packages/manager/apps/vrack-services/src/components/layout-helpers/index.ts
@@ -1,3 +1,2 @@
 export * from './DashboardLayout.component';
-export * from './PageLayout.component';
 export * from './CreatePageLayout.component';
