--- packages/manager/apps/vrack-services/src/components/layout-helpers/layout-helpers.type.ts
@@ -1,7 +1,0 @@
-export type DashboardTabItemProps = {
-  name: string;
-  title: string;
-  to: string;
-  pathMatchers?: RegExp[];
-  onClick?: () => void;
-};
