--- packages/manager/apps/vrack-services/src/utils/date.ts
@@ -1,11 +1,0 @@
-/* eslint-disable import/prefer-default-export */
-export const formatDateString = (dateString: string, locale?: string) => {
-  const date = new Date(dateString);
-  return date.toString() !== 'Invalid Date'
-    ? date.toLocaleString(locale, {
-        day: 'numeric',
-        month: 'short',
-        year: 'numeric',
-      })
-    : '-';
-};
