+++ packages/manager/apps/vrack-services/tsconfig.json
@@ -22,6 +22,6 @@
       "@/*": ["./src/*"]
     }
   },
-  "include": ["src", "useless-tests"],
+  "include": ["src"],
   "exclude": ["node_modules", "dist", "types"]
 }
